moved tribes folder into pages
