+++ frontend/app/src/pages/index.tsx
@@ -7,8 +7,8 @@ import { observer } from 'mobx-react-lite';
 import BotsBody from '../bots/Body';
 import PeopleHeader from '../people/main/Header';
 import TokenRefresh from '../people/utils/TokenRefresh';
-import Body from '../tribes/Body';
-import Header from '../tribes/Header';
+import Body from './tribes/Body';
+import Header from './tribes/Header';
 import { MainLayout } from './MainLayout';
 import { Modals } from './Modals';
 import { People } from './people';
