+++ frontend/app/src/pages/tribes/Body.tsx
@@ -9,12 +9,12 @@ import {
   EuiHighlight
 } from '@elastic/eui';
 import { observer } from 'mobx-react-lite';
-import { useStores } from '../store';
-import { useIsMobile, usePageScroll } from '../hooks';
-import { SearchTextInput } from '../components/common';
-import NoResults from '../people/utils/NoResults';
-import PageLoadSpinner from '../people/utils/PageLoadSpinner';
-import { colors } from '../config/colors';
+import { useStores } from '../../store';
+import { useIsMobile, usePageScroll } from '../../hooks';
+import { SearchTextInput } from '../../components/common';
+import NoResults from '../../people/utils/NoResults';
+import PageLoadSpinner from '../../people/utils/PageLoadSpinner';
+import { colors } from '../../config/colors';
 import tags from './tags';
 import Tag from './Tag';
 import Tribe from './Tribe';
