+++ frontend/app/src/pages/tribes/Tribe.tsx
@@ -3,7 +3,7 @@ import moment from 'moment';
 import React, { useState } from 'react';
 import { QRCode } from 'react-qr-svg';
 import styled from 'styled-components';
-import { getHostIncludingDockerHosts } from '../config/host';
+import { getHostIncludingDockerHosts } from '../../config/host';
 import Tag from './Tag';
 
 interface ContentProps {
