+++ frontend/app/src/pages/tribes/rmmd.ts
